refactor: Clean up test files by removing unused imports and consolidating utility functions
