+++ brainevent/_jitc_float_normal_impl_test.py
@@ -14,10 +14,6 @@
 # ==============================================================================
 
 
-import os
-
-os.environ['JAX_TRACEBACK_FILTERING'] = 'off'
-
 import brainstate
 import brainunit as u
 import jax
