+++ brainevent/_jitc_float_uniform_impl_test.py
@@ -14,9 +14,7 @@
 # ==============================================================================
 
 
-import os
-
-os.environ['JAX_TRACEBACK_FILTERING'] = 'off'
+from typing import Tuple
 
 import brainstate
 import brainunit as u
@@ -24,7 +22,6 @@ import jax
 import jax.numpy as jnp
 import numpy as np
 import pytest
-from typing import Tuple
 
 import brainevent
